organize page actions and add favorites option in the sidebar
