+++ components/common/PageLayout/components/Header/components/DatabasePageOptions.tsx
@@ -28,9 +28,9 @@ import {
 import { useAppSelector } from 'components/common/BoardEditor/focalboard/src/store/hooks';
 import { getCurrentBoardViews, getView } from 'components/common/BoardEditor/focalboard/src/store/views';
 import { Utils } from 'components/common/BoardEditor/focalboard/src/utils';
-import { DuplicatePageAction } from 'components/common/DuplicatePageAction';
 import type { ImportAction } from 'components/common/Modal/ConfirmImportModal';
 import ConfirmImportModal from 'components/common/Modal/ConfirmImportModal';
+import { DuplicatePageAction } from 'components/common/PageActions/components/DuplicatePageAction';
 import { useApiPageKeys } from 'hooks/useApiPageKeys';
 import { useCurrentSpace } from 'hooks/useCurrentSpace';
 import { useDateFormatter } from 'hooks/useDateFormatter';
