+++ components/common/PageLayout/components/PageNavigation/components/PageTreeItem.tsx
@@ -22,7 +22,7 @@ import charmClient from 'charmClient';
 import { getSortedBoards } from 'components/common/BoardEditor/focalboard/src/store/boards';
 import { useAppSelector } from 'components/common/BoardEditor/focalboard/src/store/hooks';
 import EmojiPicker from 'components/common/BoardEditor/focalboard/src/widgets/emojiPicker';
-import { DuplicatePageAction } from 'components/common/DuplicatePageAction';
+import { DuplicatePageAction } from 'components/common/PageActions/components/DuplicatePageAction';
 import TreeItemContent from 'components/common/TreeItemContent';
 import { useCurrentSpacePermissions } from 'hooks/useCurrentSpacePermissions';
 import { usePageFromPath } from 'hooks/usePageFromPath';
